+++ src/types.cc
@@ -33,7 +33,7 @@ SensorHit make_sensor_hit(u32 event, u16 sensor_id, f32 time) {
   return {event, sensor_id, time};
 }
 
-Interaction make_interaction(const G4ThreeVector &v, f32 e, u16 n) {
+Interaction make_interaction(const G4ThreeVector &v, f32 e, u32 n) {
   return { static_cast<f32>(v.x())
          , static_cast<f32>(v.y())
          , static_cast<f32>(v.z())
